refactor: Use BikeGoalManager instead GoalsList
- Removal of the `GoalsList` component.
- The display of the goals block is managed by `BikeGoalManager`.
- `BikeGoalSummary` manages the request on timeseries.
- `BikeGoalAlertSuccess` manages the setting `bikeGoal.showAlertSuccess`.
